adding move semantics to cache
