+++ tests/io/CMakeLists.txt
@@ -1,0 +1,6 @@
+# Sources
+set(TEST_SOURCES
+    test_io.cpp
+    test_cache.cpp
+)
+target_sources(wmtk_tests PRIVATE ${TEST_SOURCES})
